wip: refactor: also split out @fadroma/compile and @fadroma/devnet
