+++ agent/package.json
@@ -1,6 +1,6 @@
 {
   "name":    "@fadroma/agent",
-  "version": "2.0.0-rc.3",
+  "version": "2.0.0-rc.4",
   "type":    "module",
   "main":    "agent.ts",
   "files": [ "*.ts", "!*.test.*" ],
