+++ compile/compile.cli.cjs
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+// TODO
