+++ compile/compile.ts
@@ -510,3 +510,4 @@ export class ContainerizedLocalRustCompiler extends LocalRustCompiler {
   }
 
 }
+
