+++ connect/cw/package.json
@@ -6,7 +6,7 @@
   "files": [ "*.ts", "okp4/*.ts" ],
   "description": "CosmJS Stargate integration for Fadroma, incl. OKP4 support.",
   "dependencies": {
-    "@fadroma/agent":     "workspace:2.0.0-rc.3",
+    "@fadroma/agent":     "workspace:2.0.0-rc.4",
     "@hackbg/conf":       "workspace:3.1.5",
     "@hackbg/cosmjs-esm": "workspace:1.0.0-rc.32",
     "@noble/hashes":      "^1.3.2",
