+++ connect/scrt/package.json
@@ -6,7 +6,7 @@
   "files": [ "*.ts" ],
   "description": "Fadroma support for Secret Network. Provides a SecretJS-based implementation of the base classes in @fadroma/agent to operate smart contracts on the Secret Network.",
   "dependencies": {
-    "@fadroma/agent":       "workspace:2.0.0-rc.3",
+    "@fadroma/agent":       "workspace:2.0.0-rc.4",
     "@hackbg/conf":         "workspace:3.1.5",
     "@hackbg/secretjs-esm": "workspace:1.9.3-patch.29",
     "@noble/secp256k1":     "^2.0.0",
