+++ devnet/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@fadroma/devnet",
+  "version": "0.1.0",
+  "main": "project.ts",
+  "dependencies": {
+    "@hackbg/file": "workspace:*",
+    "@hackbg/logs": "workspace:*",
+    "@hackbg/cmds": "workspace:*",
+    "@fadroma/agent": "workspace:*"
+  }
+}
