+++ package.json
@@ -53,7 +53,7 @@
     "release:fast": "pnpm i && ubik release --otp 123123"
   },
   "dependencies": {
-    "@fadroma/agent": "workspace:2.0.0-rc.3",
+    "@fadroma/agent": "workspace:2.0.0-rc.4",
     "@fadroma/connect": "workspace:4.0.0-rc.2",
     "@hackbg/cmds": "workspace:*",
     "@hackbg/conf": "workspace:*",
